requirements.txt
# PyTorch Core & Vision
torch>=1.10.0
torchvision>=0.11.0

numpy>=1.21.0
scikit-learn>=1.0.0

PyYAML>=6.0

matplotlib>=3.4.0